% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{sigma}
\alias{sigma}
\title{Residual variance components}
\usage{
sigma(object, ...)
}
\arguments{
\item{object}{object of class \code{\link{clme}}.}

\item{...}{space for additional arguments}
}
\description{
Extract residual variance components for objects of class \code{clme}.
}

