% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{is.clme}
\alias{as.clme}
\alias{is.clme}
\title{Constructor method for objects S3 class clme}
\usage{
is.clme(x)

as.clme(x, ...)
}
\arguments{
\item{x}{list with the elements corresponding to the output of \code{\link{clme}}.}

\item{...}{space for additional arguments.}
}
\value{
Returns an object of the class \code{clme}.
}
\description{
Constructor method for objects S3 class clme

Test if an object is of class \code{clme} or coerce an object to be such.
}
\examples{
data( rat.blood )

cons <- list(order = "simple", decreasing = FALSE, node = 1 )
clme.out <- clme(mcv ~ time + temp + sex + (1|id), data = rat.blood , 
                 constraints = cons, seed = 42, nsim = 0)

is.clme( clme.out )
as.clme( clme.out )
}
\seealso{
\code{\link{CLME-package}}
\code{\link{clme}}
}

