% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{fixef.clme}
\alias{coef.clme}
\alias{coef.summary.clme}
\alias{coefficients.clme}
\alias{coefficients.summary.clme}
\alias{fixed.effects}
\alias{fixed.effects.clme}
\alias{fixed.effects.summary.clme}
\alias{fixef.clme}
\alias{fixef.summary.clme}
\title{Extract fixed effects}
\usage{
\method{fixef}{clme}(object, ...)

\method{fixef}{summary.clme}(object, ...)

\method{fixef}{clme}(object, ...)

fixed.effects(object, ...)

\method{fixed.effects}{summary.clme}(object, ...)

\method{fixed.effects}{clme}(object, ...)

\method{coefficients}{clme}(object, ...)

\method{coef}{clme}(object, ...)

\method{coefficients}{summary.clme}(object, ...)

\method{coef}{summary.clme}(object, ...)
}
\arguments{
\item{object}{object of class \code{\link{clme}}.}

\item{...}{space for additional arguments}
}
\value{
Returns a numeric vector.
}
\description{
Extract fixed effects

Extract fixed effects

Extracts the fixed effects estimates from objects of class \code{clme}.

Extract fixed effects

Extract fixed effects
}
\examples{
data( rat.blood )
cons <- list(order = "simple", decreasing = FALSE, node = 1 )
clme.out <- clme(mcv ~ time + temp + sex + (1|id), data = rat.blood ,
                 constraints = cons, seed = 42, nsim = 0)

fixef( clme.out )
}
\seealso{
\code{\link{CLME-package}}
\code{\link{clme}}
}

