% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{model.frame.clme}
\alias{model.frame.clme}
\alias{model.frame.summary.clme}
\title{Extracts the model frame}
\usage{
\method{model.frame}{clme}(formula, ...)

\method{model.frame}{summary.clme}(formula, ...)
}
\arguments{
\item{formula}{a formula expression.}

\item{...}{space for additional arguments}
}
\value{
Returns a data frame with the variables in the model.
}
\description{
Extracts the model frame from objects of class \code{clme}.

Extracts the model frame
}
\examples{
\dontrun{
data( rat.blood )
model.frame.clme( mcv ~ time + temp + sex + (1|id), data = rat.blood )
}
}
\seealso{
\code{\link{CLME-package}}
\code{\link{clme}}

\code{\link{model.frame.clme}}
}

