% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{model.matrix.clme}
\alias{model.matrix.clme}
\title{Extract the fixed-effects design matrix.}
\usage{
\method{model.matrix}{clme}(object, type = "fixef", ...)
}
\arguments{
\item{object}{an object of class \code{clme}.}

\item{type}{specify whether to return the fixed-effects or random-effects matrix.}

\item{...}{space for additional arguments}
}
\value{
Returns a matrix.
}
\description{
Extracts the fixed-effects design matrix from objects of class \code{clme}.
}
\examples{
\dontrun{
data( rat.blood )
cons <- list(order = "simple", decreasing = FALSE, node = 1 )
clme.out <- clme(mcv ~ time + temp + sex + (1|id), data = rat.blood ,
                 constraints = cons, seed = 42, nsim = 0)

model.matrix( clme.out )
}
}
\seealso{
\code{\link{CLME-package}}
\code{\link{clme}}
}

