% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{ranef.clme}
\alias{random.effects}
\alias{ranef.clme}
\title{Extract random effects}
\usage{
ranef.clme(object, ...)

random.effects(object, ...)
}
\arguments{
\item{object}{object of class clme.}

\item{...}{space for additional arguments}
}
\description{
Extract random effects

Extract random effects
}

