% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{random.effects}
\alias{random.effects}
\alias{random.effects.summary.clme}
\title{Extract random effects}
\usage{
random.effects(object, ...)

\method{random.effects}{summary.clme}(object, ...)
}
\arguments{
\item{object}{object of class clme.}

\item{...}{space for additional arguments}
}
\description{
Extract random effects

Extract random effects
}

