% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{random.effects}
\alias{random.effects}
\alias{random.effects.summary.clme}
\title{Extract random effects}
\usage{
random.effects(object, ...)

\method{random.effects}{summary.clme}(object, ...)
}
\description{
Extract random effects

Extract random effects
}

