% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_clme.r
\docType{data}
\name{shiny_clme}
\alias{shinyServer_clme}
\alias{shinyUI_clme}
\alias{shiny_clme}
\title{Shiny GUI for CLME}
\format{\preformatted{List of 3
 $ : NULL
 $ : NULL
 $ :List of 1
  ..$ :List of 3
  .. ..$ name    : chr "div"
  .. ..$ attribs :List of 1
  .. .. ..$ class: chr "container-fluid"
  .. ..$ children:List of 2
  .. .. ..$ :List of 2
  .. .. .. ..$ :List of 3
  .. .. .. .. ..$ name    : chr "head"
  .. .. .. .. ..$ attribs : list()
  .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. ..$ name    : chr "title"
  .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. ..$ : chr "Constrained Linear Mixed Effects"
  .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..$ :List of 3
  .. .. .. .. ..$ name    : chr "h2"
  .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. ..$ style: chr "padding: 10px 0px;"
  .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. ..$ : chr "Constrained Linear Mixed Effects"
  .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. ..$ :List of 3
  .. .. .. ..$ name    : chr "div"
  .. .. .. ..$ attribs :List of 1
  .. .. .. .. ..$ class: chr "row-fluid"
  .. .. .. ..$ children:List of 2
  .. .. .. .. ..$ :List of 3
  .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. ..$ class: chr "span4"
  .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. ..$ name    : chr "form"
  .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. ..$ class: chr "well"
  .. .. .. .. .. .. .. ..$ children:List of 35
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Data (choose CSV file)"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ id    : chr "file1"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type  : chr "file"
  .. .. .. .. .. .. .. .. .. .. .. ..$ accept: chr "text/csv,text/comma-separated-values,text/plain,.csv"
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "file1_progress"
  .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "progress progress-striped active shiny-file-input-progress"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "bar"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "dlmt"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Delimiter:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "dlmt"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :Classes 'html', 'character'  atomic [1:1] <option value="Comma" selected>Comma</option>
<option value="Tab">Tab</option>
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "dlmt"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "\{\}"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 7
  .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.8.5"
  .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap2.css"
  .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\\n<script src=\\"shared/selectize/js/es5-shim.min.js\\"></script>\\n<![endif]-->\\n<script src=\\"shared/selectize/"| __truncated__
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "solver"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Type of Solver / fitting norm:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "solver"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :Classes 'html', 'character'  atomic [1:1] <option value="Least Squares (LS)" selected>Least Squares (LS)</option>
<option value="Least Absolute Value (L1)">Least Absolute Value (L1)</option>
<option value="General LS (GLS)">General LS (GLS)</option>
<option value="Asymmetrix LS">Asymmetrix LS</option>
<option value="L1 approx">L1 approx</option>
<option value="Huber">Huber</option>
<option value="SILF">SILF</option>
<option value="Chebyshev">Chebyshev</option>
<option value="L-p Power Norm">L-p Power Norm</option>
<option value="Quantile">Quantile</option>
<option value="Poisson">Poisson</option>
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "solver"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "\{\}"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 7
  .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.8.5"
  .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap2.css"
  .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\\n<script src=\\"shared/selectize/js/es5-shim.min.js\\"></script>\\n<![endif]-->\\n<script src=\\"shared/selectize/"| __truncated__
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "tsfunc"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Test Statistic:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "tsfunc"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :Classes 'html', 'character'  atomic [1:1] <option value="LRT" selected>LRT</option>
<option value="Williams">Williams</option>
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "tsfunc"
  .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "\{\}"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 7
  .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.8.5"
  .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap2.css"
  .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\\n<script src=\\"shared/selectize/js/es5-shim.min.js\\"></script>\\n<![endif]-->\\n<script src=\\"shared/selectize/"| __truncated__
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "order"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-group shiny-input-checkboxgroup"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "order"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Order (select at least one):"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox "
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 4
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name : chr "order"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "order1"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "Simple"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Simple"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox "
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 4
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name : chr "order"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "order2"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "Umbrella"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Umbrella"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox "
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 4
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name : chr "order"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "order3"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "Tree"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Tree"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "Identify columns of data"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "Use column letters or numbers"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "e.g., 1-3 or A-C or a list: 1,4,6"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "yy"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column of response:"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "yy"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr ""
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "p1"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column of constrained effect:"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "p1"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr ""
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "p2"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column(s) of Covariates:"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "p2"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "None"
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "q"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column(s) of random effects:"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "q"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "None"
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "decreasing"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "decreasing"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Decreasing order (inverted):"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "nsim"
  .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Number of Bootstraps:"
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "nsim"
  .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "1000"
  .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 50000
  .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "Click to run model or update output."
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "compute1"
  .. .. .. .. .. .. .. .. .. .. ..$ type : chr "button"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "btn action-button"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Run model"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "compute2"
  .. .. .. .. .. .. .. .. .. .. ..$ type : chr "button"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "btn action-button"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Update plot"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "compute3"
  .. .. .. .. .. .. .. .. .. .. ..$ type : chr "button"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "btn action-button"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Update summary"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "outcheck"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "outcheck"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Format Output:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.outcheck"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "plotci"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "plotci"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "CI on Plot:"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.outcheck"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "alpha"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Alpha level:"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 14
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id         : chr "alpha"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type       : chr "slider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name       : chr "alpha"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value      : chr "0.05"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class      : chr "jslider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-from  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-to    : num 0.15
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-step  : num 0.01
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-skin  : chr "plastic"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-round : logi FALSE
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-locale: chr "us"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-format: chr "#,##0.#####"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-scale : chr "|;|;|;|;|;|;|;|;|;|;|;|;|;|;|;|"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-smooth: logi FALSE
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 7
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "jslider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "1"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/slider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "js/jquery.slider.min.js"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/jquery.slider.min.css"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "varssq"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "varssq"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Heteroscedasticity:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.varssq"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "gfix"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column of variance groups:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "gfix"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr ""
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "xlevel1"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "xlevel1"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Define order of constrained groups:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.xlevel1"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "xlevels"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Column of ordered group levels:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "xlevels"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr ""
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "technical"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Select Control Parameters:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "emiter"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Max EM Iterations:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "emiter"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "500"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 10
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 50000
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "mqiter"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Max MINQUE Iterations:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "mqiter"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "500"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 10
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 50000
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "emeps"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "EM Convergence Criteria:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "emeps"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "0.0001"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 50000
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "mqeps"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "MINQUE Convergence Criteria:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "mqeps"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "0.0001"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 50000
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.technical"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "ranseed"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Set RNG Seed:"
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 5
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "ranseed"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "42"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num Inf
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. ..$ :List of 3
  .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. ..$ class: chr "span8"
  .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. ..$ class: chr "tabbable tabs-above"
  .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "ul"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "nav nav-tabs"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "li"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "active"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "a"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href       : chr "#tab-9218-1"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-toggle: chr "tab"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Data Summary"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "li"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "a"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href       : chr "#tab-9218-2"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-toggle: chr "tab"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Model Summary"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "li"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "a"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href       : chr "#tab-9218-3"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-toggle: chr "tab"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Model Plot"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "tab-content"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "tab-pane active"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "tab-9218-1"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "fig0"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "shiny-plot-output"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ style: chr "width: 100\% ; height: 650px"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "sum_table"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "shiny-html-output"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "tab-pane"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "tab-9218-2"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "pre"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "summary"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "shiny-text-output"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "tab-pane"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "tab-9218-3"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "fig1"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "shiny-plot-output"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ style: chr "width: 100\% ; height: 650px"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. ..- attr(*, "class")= chr "shiny.tag"
 - attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
 - attr(*, "html_dependencies")=List of 2
  ..$ :List of 7
  .. ..$ name      : chr "bootstrap"
  .. ..$ version   : chr "2.3.2"
  .. ..$ src       :List of 2
  .. .. ..$ href: chr "shared/bootstrap"
  .. .. ..$ file: chr "/home/statman/R/x86_64-pc-linux-gnu-library/3.0/shiny/www/shared/bootstrap"
  .. ..$ meta      : NULL
  .. ..$ script    : chr "js/bootstrap.min.js"
  .. ..$ stylesheet: chr "css/bootstrap.min.css"
  .. ..$ head      : NULL
  .. ..- attr(*, "class")= chr "html_dependency"
  ..$ :List of 7
  .. ..$ name      : chr "bootstrap-responsive"
  .. ..$ version   : chr "2.3.2"
  .. ..$ src       :List of 2
  .. .. ..$ href: chr "shared/bootstrap"
  .. .. ..$ file: chr "/home/statman/R/x86_64-pc-linux-gnu-library/3.0/shiny/www/shared/bootstrap"
  .. ..$ meta      :List of 1
  .. .. ..$ viewport: chr "width=device-width, initial-scale=1.0"
  .. ..$ script    : NULL
  .. ..$ stylesheet: chr "css/bootstrap-responsive.min.css"
  .. ..$ head      : NULL
  .. ..- attr(*, "class")= chr "html_dependency"
}}
\usage{
shiny_clme()

shinyUI_clme

shinyServer_clme(input, output)
}
\arguments{
\item{input}{input from GUI.}

\item{output}{output to GUI.}
}
\description{
Opens a graphical user interface to run \pkg{CLME}, built from the \pkg{shiny} package.




}
\details{
Currently the GUI does not allow specification of custom orders for the alternative hypothesis. Future versions may enable this capability.
The data should be a CSV or table-delimited file with the first row being a header. Variables are identified using their column letter or number (e.g., 1 or A). Separate multiple variables with a comma (e.g., 1,2,4 or A,B,D), or select a range of variables with a dash (e.g., 1-4 or A-D). Set to 'None' (default) to indicate no covariates or random effects.
If group levels for the constrained effect are character, they may not be read in the proper order. An extra column may contain the ordered group levels (it may therefore have different length than the rest of the dataset).
}
\note{
This function is primarily designed to be called by \code{\link{clme}}.

By default, homogeneous variances are assumed for the residuals and (if included) random effects. Heterogeneity can be induced using the arguments \code{Nks} and \code{Qs}, which refer to the vectors \eqn{ (n_{1}, n_{2}, \ldots, n_{k}) }{(n1, n2 ,... , nk)} and \eqn{ (c_{1}, c_{2}, \ldots, c_{q}) }{(c1, c2 ,... , cq)}, respectively. See \code{\link{clme_em}} for further explanation of these values.
}
\examples{
\dontrun{ shiny_clme() }
}
\keyword{datasets}

