% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_clme.r
\docType{data}
\name{shiny_clme}
\alias{shiny_clme}
\alias{shinyUI_clme}
\alias{shinyServer_clme}
\title{Shiny GUI for CLME}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
shiny_clme()

shinyUI_clme

shinyServer_clme(input, output)
}
\arguments{
\item{input}{input from GUI.}

\item{output}{output to GUI.}
}
\description{
Opens a graphical user interface to run \pkg{CLME}, built from the \pkg{shiny} package.

The UI for the shiny app in CLME

The server for the shiny app in CLME
}
\details{
Currently the GUI does not allow specification of custom orders for the alternative hypothesis. Future versions may enable this capability.
The data should be a CSV or table-delimited file with the first row being a header. Variables are identified using their column letter or number (e.g., 1 or A). Separate multiple variables with a comma (e.g., 1,2,4 or A,B,D), or select a range of variables with a dash (e.g., 1-4 or A-D). Set to 'None' (default) to indicate no covariates or random effects.
If group levels for the constrained effect are character, they may not be read in the proper order. An extra column may contain the ordered group levels (it may therefore have different length than the rest of the dataset).
}
\note{
This function is primarily designed to call \code{\link{clme}}.
}
\examples{
\dontrun{ shiny_clme() }

}
\keyword{datasets}
