% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{sigma.clme}
\alias{sigma.clme}
\title{Residual variance components}
\usage{
sigma.clme(object, ...)
}
\arguments{
\item{object}{object of class clme.}

\item{...}{space for additional arguments}
}
\description{
Residual variance components
}

